refactor: payroll_model
